refactor(blorktools): util rename
