+++ packages/blorktools/src/asset_debugger/axis-indicator/axis-indicator.html
