+++ packages/blorktools/src/asset_debugger/landing-page/file-handler.js
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import { getState, updateState } from '../scene/state';
+import { getState, updateState } from '../util/state/scene-state';
 import { updateAtlasVisualization } from '../panels/atlas-panel/atlas-panel.js';
 import * as worldPanelModule from '../panels/world-panel/world-panel.js';
 import { 
@@ -8,9 +8,9 @@ import {
   processLightingFile, 
   terminateAllWorkers 
 } from '../util/workers/worker-manager.js';
-import { parseLightingData } from '../util/scene/lighting-util';
-import { clearDropzone, setupDropzone } from '../util/upload/file-upload-util';
-import { createGLBPreview } from '../util/upload/glb-preview-util';
+import { parseLightingData } from '../util/scene/lighting-manager';
+import { clearDropzone, setupDropzone } from '../util/upload/file-upload-handler';
+import { createGLBPreview } from '../util/upload/glb-preview-controller';
 
 // Add event listener to terminate all workers when the page is unloaded
 window.addEventListener('beforeunload', () => {
