+++ packages/blorktools/src/asset_debugger/modals/settings-modal/settings-modal.js
@@ -1,7 +1,7 @@
 /**
  * Settings Modal UI Component for Asset Debugger
  */
-import { saveSettings, loadSettings, getDefaultSettings } from '../../util/data/localstorage-util.js';
+import { saveSettings, loadSettings, getDefaultSettings } from '../../util/data/localstorage-manager.js';
 import { updateRigOptions } from '../../util/scene/rig/rig-controller.js';
 
 export class SettingsModal {
