+++ packages/blorktools/src/asset_debugger/panels/world-panel/world-panel.js
@@ -3,8 +3,8 @@
  * 
  * This module handles world properties, environment, and lighting visualization and controls.
  */
-import { getState } from '../../scene/state';
-import { updateLighting, resetLighting, updateExposure } from '../../util/scene/lighting-util';
+import { getState } from '../../util/state/scene-state';
+import { updateLighting, resetLighting, updateExposure } from '../../util/scene/lighting-manager';
 import * as THREE from 'three';
 
 // Track initialization state
@@ -539,7 +539,7 @@ function setupBgToggleListener() {
                             console.log('Attempting to reload environment texture from lighting file');
                             
                             // Import lighting utility and load the texture again
-                            import('../../util/scene/lighting-util.js').then(lightingModule => {
+                            import('../../util/scene/lighting-manager.js').then(lightingModule => {
                                 if (lightingModule.setupEnvironmentLighting && state.lightingFile) {
                                     lightingModule.setupEnvironmentLighting(state.lightingFile)
                                         .then(newTexture => {
@@ -1684,7 +1684,7 @@ export function generatePreviewOnly(file, previewElement) {
                     backgroundTextureCache.set(file.name, texture);
                     
                     // Also store in state for use with debugging
-                    import('../../scene/state.js').then(stateModule => {
+                    import('../../util/state/scene-state.js').then(stateModule => {
                         stateModule.updateState({
                             backgroundTexture: texture
                         });
