+++ packages/blorktools/src/asset_debugger/util/animation/playback/animation-playback-controller.js
@@ -1,4 +1,4 @@
-import { frameInterval } from "../../upload/animation-preview-util";
+import { frameInterval } from "../../upload/animation-preview-controller";
 import { 
     ANALYSIS_DURATION_MS,
     animationDuration,
