+++ packages/blorktools/src/asset_debugger/util/data/glb-binary-buffer-handler.js
