+++ packages/blorktools/src/asset_debugger/util/data/localstorage-manager.js
