+++ packages/blorktools/src/asset_debugger/util/scene/camera-controller.js
@@ -7,7 +7,7 @@
  */
 import * as THREE from 'three';
 import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
-import { getState, updateState } from '../../scene/state';
+import { getState, updateState } from '../state/scene-state';
 
 // Store a reference to the controls instance
 let controlsInstance = null;
