+++ packages/blorktools/src/asset_debugger/util/scene/model-handler.js
@@ -6,9 +6,9 @@
 
 import * as THREE from 'three';
 import { GLTFLoader } from 'three/addons/loaders/GLTFLoader.js';
-import { getState, updateState } from '../../scene/state.js';
-import { createMaterial } from './material-creation-util.js';
-import { fitCameraToObject } from '../../scene/scene.js';
+import { getState, updateState } from '../state/scene-state.js';
+import { createMaterial } from './pbr-material-factory.js';
+import { fitCameraToObject } from './threejs-scene-controller.js';
 import { createMeshVisibilityPanel } from '../../panels/mesh-panel/mesh-panel.js';
 import { updateAtlasVisualization } from '../../panels/atlas-panel/atlas-panel.js';
 import { updateUvPanel } from '../../panels/uv-panel/uv-panel.js';
@@ -30,7 +30,7 @@ export function loadAndSetupModel(loadingIndicator) {
             
             try {
                 Promise.all([
-                    import('../data/mesh-html-util.js'),
+                    import('../data/mesh-html-manager.js'),
                     import('../../modals/html-editor-modal/model-integration.js')
                 ]).then(([meshDataUtil, modelIntegration]) => {
                     let bufferPromises = [];
@@ -356,7 +356,7 @@ function prepareGlbBuffer() {
     
     if (state.useCustomModel && state.modelFile) {
         return Promise.all([
-            import('../data/mesh-html-util.js'),
+            import('../data/mesh-html-manager.js'),
             import('../../modals/html-editor-modal/model-integration.js')
         ]).then(([meshDataUtil, modelIntegration]) => {
             if (modelIntegration.processModelFileForHtmlEditor) {
