+++ packages/blorktools/src/asset_debugger/util/scene/pbr-material-factory.js
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import { getState, updateState } from '../../scene/state.js';
+import { getState, updateState } from '../state/scene-state.js';
 
 /**
  * Create a PBR material with the loaded textures
