+++ packages/blorktools/src/asset_debugger/util/scene/threejs-preview-setup.js
@@ -12,9 +12,9 @@ import {
 } from "../state/threejs-state";
 import { showStatus } from '../../modals/html-editor-modal/html-editor-modal';
 import { createMeshInfoPanel } from '../../modals/html-editor-modal/mesh-info-panel-util';
-import { createTextureFromIframe } from '../animation/render/iframe2texture-render';
-import { getState } from '../../scene/state';
-import { animatePreview } from '../upload/animation-preview-util';
+import { createTextureFromIframe } from '../animation/render/iframe2texture-render-controller';
+import { getState } from '../state/scene-state';
+import { animatePreview } from '../upload/animation-preview-controller';
 import { setIsPreviewActive } from '../state/animation-state';
 
 /**
