+++ packages/blorktools/src/asset_debugger/util/scene/threejs-scene-controller.js
@@ -4,10 +4,10 @@
  * This module handles Three.js scene setup, rendering, and animation.
  */
 import * as THREE from 'three';
-import { getState, updateState } from './state.js';
-import { updateRigAnimation } from '../util/scene/rig/rig-controller.js';
-import { addLighting, setupEnvironmentLighting } from '../util/scene/lighting-util.js';
-import { createControls, updateControls, setControlsTarget } from '../util/scene/controls-util.js';
+import { getState, updateState } from '../state/scene-state.js';
+import { updateRigAnimation } from './rig/rig-controller.js';
+import { addLighting, setupEnvironmentLighting } from './lighting-manager.js';
+import { createControls, updateControls, setControlsTarget } from './camera-controller.js';
 
 /**
  * Initialize the Three.js scene, camera, renderer and controls
