+++ packages/blorktools/src/asset_debugger/util/scene/ui-manager.js
@@ -3,7 +3,7 @@
  * 
  * This module handles the UI interactions and tab switching.
  */
-import { getState } from '../../scene/state.js';
+import { getState } from '../state/scene-state.js';
 import { updateAtlasVisualization } from '../../panels/atlas-panel/atlas-panel.js';
 import { updateUvPanel } from '../../panels/uv-panel/uv-panel.js';
